Cleaned up some read/write functions and files. Added ability to read and write binary files with Codable
