+++ Treviz/Model/Condition.swift
@@ -403,7 +403,7 @@ class Condition : NSObject, EvaluateCondition, Codable {
         } catch { _summary = "" }
     }
     
-    func encode(to encoder: Encoder) throws {
+    func encode(to encoder: Encoder) throws { 
         var container = encoder.container(keyedBy: CodingKeys.self)
         try container.encode(name, forKey: .name)
         try container.encode(unionType, forKey: .unionType)
